requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
requests==2.32.3
SQLAlchemy==2.0.35
translate==3.6.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.1.2
